% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyview.R
\name{shiny_view}
\alias{shiny_view}
\title{Display a data.frame through a Shiny application.}
\usage{
shiny_view(df)
}
\arguments{
\item{df}{data.frame. The data to display.}

\item{open}{logical. Whether to automatically open the web page view system call.}
}
\description{
Display a data.frame through a Shiny application.
}
\examples{
\dontrun{
  shiny_view(iris) # Will open a nice web page to browse through the iris dataset.
}
}
